requirements.txt
Flask_Uploads==0.2.1
Flask_Bootstrap==3.3.7.1
Flask_WTF==0.14.2
Flask_Mail==0.9.1
Markdown==2.6.8
tqdm==4.59.0
Flask_HTTPAuth==3.2.3
Flask_SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Flask_Migrate==2.0.4
Flask_Script==2.0.6
itsdangerous==0.24
Flask_Login==0.4.0
bleach==2.0.0
Faker==8.11.0
WTForms==2.1
Werkzeug==0.12.2
Flask_Moment==0.5.1
Flask_PageDown==0.2.2
numpy==1.19.3
Flask==0.12.2
alembic==0.9.3
Flask_SSLify==0.1.5
Pillow==8.3.2
pyOpenSSL==20.0.1